+++ android/src/main/kotlin/dev/msfjarvis/claw/android/viewmodel/DataTransferRepository.kt
@@ -13,6 +13,8 @@ import dev.msfjarvis.claw.database.local.SavedPost
 import dev.msfjarvis.claw.database.local.SavedPostQueries
 import java.io.InputStream
 import java.io.OutputStream
+import java.time.Instant
+import java.time.format.DateTimeFormatter
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
@@ -49,5 +51,54 @@ constructor(
   suspend fun exportPosts(output: OutputStream) {
     val posts = withContext(dbDispatcher) { savedPostQueries.selectAllPosts().executeAsList() }
     withContext(ioDispatcher) { json.encodeToStream(serializer, posts, output) }
+  }
+
+  suspend fun exportPostsAsHTML(output: OutputStream) {
+    fun computeTimestamp(post: SavedPost): Long {
+      val temporal = DateTimeFormatter.ISO_OFFSET_DATE_TIME.parse(post.createdAt)
+      val instant = Instant.from(temporal)
+      return instant.toEpochMilli()
+    }
+
+    val posts = withContext(dbDispatcher) { savedPostQueries.selectAllPosts().executeAsList() }
+    val header =
+      """
+      <!DOCTYPE NETSCAPE-Bookmark-file-1>
+      <!-- This is an automatically generated file.
+           It will be read and overwritten.
+           DO NOT EDIT! -->
+      <META HTTP-EQUIV="Content-Type" CONTENT="text/html; charset=UTF-8">
+      <TITLE>Bookmarks</TITLE>
+      <H1>Bookmarks</H1>
+    """
+        .trimIndent()
+    val html = buildString {
+      append(header)
+      append("<DD><p>\n")
+      for (post in posts) {
+        append(
+          """
+          <DT><A HREF="${post.url.ifEmpty { post.commentsUrl }}" ADD_DATE="${computeTimestamp(post)}" PRIVATE="0" TAGS="${post.tags.joinToString(",")}">${post.title}</A>
+          <DD>${post.title}
+
+        """
+            .trimIndent()
+        )
+      }
+      append(
+        """
+        <DT><A HREF="https://example.com/" ADD_DATE="0" PRIVATE="0" TAGS="delete,me,pls">Padding post</A>
+        <DD>Linkding ignores the last entry so this pads the difference for imports
+
+      """
+          .trimIndent()
+      )
+      append("</DD></p>\n")
+    }
+    withContext(ioDispatcher) {
+      val writer = output.bufferedWriter()
+      writer.write(html)
+      writer.flush()
+    }
   }
 }